move OlMeasurementHandler to own package
